requirements.txt
feedparser==6.0.11
pyyaml==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
anthropic==0.39.0
